functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.5.0,<7.0.0
python-dotenv>=1.0.1,<2.0.0
